+++ src/main/java/com/example/skooldio/controller/BasketController.java
@@ -1,5 +1,6 @@
 package com.example.skooldio.controller;
 
+import com.example.skooldio.config.security.AuthUtil;
 import com.example.skooldio.entity.Basket;
 import com.example.skooldio.model.request.BasketExceptUserIdModel;
 import com.example.skooldio.model.response.BasketResponseModel;
@@ -34,14 +35,19 @@ public class BasketController {
 
     @PostMapping
     @ApiOperation(value = "create basket", response = Basket.class)
-    public ResponseModel<Basket> create(@RequestBody BasketRequestModel model) {
+    public ResponseModel<Basket> create(@RequestHeader String jwtToken, @RequestBody BasketRequestModel model) {
         ResponseModel<Basket> responseModel = new ResponseModel<>();
         try {
-            Basket basket = service.create(model);
-            responseModel.setData(basket);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Basket basket = service.create(model);
+                responseModel.setData(basket);
 
-            responseModel.setMsg("Success");
-            responseModel.setErrorMsg(null);
+                responseModel.setMsg("Success");
+                responseModel.setErrorMsg(null);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -52,12 +58,17 @@ public class BasketController {
 
     @PutMapping(CHECKOUT_ENDPOINT + "/{userId}")
     @ApiOperation(value = "checkout basket by userid", response = Basket.class)
-    public ResponseModel<BasketResponseModel> checkout(@PathVariable Long userId, @RequestBody List<Integer> productsIdList) {
+    public ResponseModel<BasketResponseModel> checkout(@RequestHeader String jwtToken, @PathVariable Long userId, @RequestBody List<Integer> productsIdList) {
         ResponseModel<BasketResponseModel> responseModel = new ResponseModel<>();
         try {
-            BasketResponseModel basketResponseModel = service.checkout(userId, productsIdList);
-            responseModel.setMsg("Success");
-            responseModel.setData(basketResponseModel);
+            if (AuthUtil.validateToken(jwtToken)) {
+                BasketResponseModel basketResponseModel = service.checkout(userId, productsIdList);
+                responseModel.setMsg("Success");
+                responseModel.setData(basketResponseModel);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             ex.printStackTrace();
             responseModel.setMsg("Failed");
@@ -68,15 +79,20 @@ public class BasketController {
 
     @PatchMapping("/{id}")
     @ApiOperation(value = "update basket by id every field except userId", response = Basket.class)
-    public ResponseModel<Basket> updateExceptUsername(@PathVariable Long id, @RequestBody BasketExceptUserIdModel model) {
+    public ResponseModel<Basket> updateExceptUsername(@RequestHeader String jwtToken, @PathVariable Long id, @RequestBody BasketExceptUserIdModel model) {
         ResponseModel<Basket> responseModel = new ResponseModel<>();
         try {
-            Basket basket = service.updateExceptUserId(id, model);
-            responseModel.setMsg("Success");
-            responseModel.setData(basket);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Basket basket = service.updateExceptUserId(id, model);
+                responseModel.setMsg("Success");
+                responseModel.setData(basket);
 
-            if (null == basket) {
-                responseModel.setErrorMsg(String.format("basket id %d not found", id));
+                if (null == basket) {
+                    responseModel.setErrorMsg(String.format("basket id %d not found", id));
+                }
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
             }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
@@ -87,18 +103,22 @@ public class BasketController {
 
     @GetMapping("/{id}")
     @ApiOperation(value = "get basket by id", response = Basket.class)
-    public ResponseModel<Basket> getById(@PathVariable Long id) {
+    public ResponseModel<Basket> getById(@RequestHeader String jwtToken, @PathVariable Long id) {
         ResponseModel<Basket> responseModel = new ResponseModel<>();
         try {
-            Basket basket = service.getById(id);
-            responseModel.setData(basket);
-            responseModel.setMsg("Success");
-            responseModel.setErrorMsg(null);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Basket basket = service.getById(id);
+                responseModel.setData(basket);
+                responseModel.setMsg("Success");
+                responseModel.setErrorMsg(null);
 
-            if (null == basket) {
-                responseModel.setErrorMsg(String.format("basket id %d not found", id));
+                if (null == basket) {
+                    responseModel.setErrorMsg(String.format("basket id %d not found", id));
+                }
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
             }
-
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -109,6 +129,7 @@ public class BasketController {
     @GetMapping(LIST_BY_USERID_ENDPOINT + "/{userId}")
     @ApiOperation(value = "get basket list by userid", response = Basket.class)
     public ResponseModel<BasketResponseModel> listByUserId(
+            @RequestHeader String jwtToken,
             @PathVariable Long userId,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size,
@@ -116,11 +137,15 @@ public class BasketController {
             @RequestParam(defaultValue = "asc") String dir) {
         ResponseModel<BasketResponseModel> responseModel = new ResponseModel<>();
         try {
-            BasketResponseModel baskets = service.listByUserId(userId, page, size, sort, dir);
-            responseModel.setData(baskets);
-            responseModel.setMsg("Success");
-            responseModel.setErrorMsg(null);
-
+            if (AuthUtil.validateToken(jwtToken)) {
+                BasketResponseModel baskets = service.listByUserId(userId, page, size, sort, dir);
+                responseModel.setData(baskets);
+                responseModel.setMsg("Success");
+                responseModel.setErrorMsg(null);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -130,14 +155,18 @@ public class BasketController {
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "delete basket by id", response = Basket.class)
-    public ResponseModel<Basket> delete(@PathVariable Long id) {
+    public ResponseModel<Basket> delete(@RequestHeader String jwtToken, @PathVariable Long id) {
         ResponseModel<Basket> responseModel = new ResponseModel<>();
         try {
-            service.deleteById(id);
-            responseModel.setData(null);
-            responseModel.setMsg(String.format("user id %d deleted.", id));
-            responseModel.setErrorMsg(null);
-
+            if (AuthUtil.validateToken(jwtToken)) {
+                service.deleteById(id);
+                responseModel.setData(null);
+                responseModel.setMsg(String.format("user id %d deleted.", id));
+                responseModel.setErrorMsg(null);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -148,6 +177,7 @@ public class BasketController {
     @GetMapping
     @ApiOperation(value = "get basket as list", response = Basket.class)
     public ResponseListModel<Basket> listPaging(
+            @RequestHeader String jwtToken,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size,
             @RequestParam(defaultValue = "id") String sort,
@@ -155,24 +185,29 @@ public class BasketController {
     ) {
         ResponseListModel<Basket> responseListModel = new ResponseListModel<>();
         try {
-            List<Basket> baskets = service.listPaging(page, size, sort, dir);
+            if (AuthUtil.validateToken(jwtToken)) {
+                List<Basket> baskets = service.listPaging(page, size, sort, dir);
 
-            int countAll = service.countAll();
-            int count = baskets.size();
-            int next = 0;
-            if (count >= size) {
-                next = page + size;
-            }
-            if (next >= countAll) {
-                next = 0;
-            }
+                int countAll = service.countAll();
+                int count = baskets.size();
+                int next = 0;
+                if (count >= size) {
+                    next = page + size;
+                }
+                if (next >= countAll) {
+                    next = 0;
+                }
 
-            responseListModel.setDatas(baskets);
-            responseListModel.setAll(countAll);
-            responseListModel.setCount(count);
-            responseListModel.setNext(next);
-            responseListModel.setMsg("Success");
-            responseListModel.setErrorMsg(null);
+                responseListModel.setDatas(baskets);
+                responseListModel.setAll(countAll);
+                responseListModel.setCount(count);
+                responseListModel.setNext(next);
+                responseListModel.setMsg("Success");
+                responseListModel.setErrorMsg(null);
+            } else {
+                responseListModel.setMsg("Failed");
+                responseListModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseListModel.setMsg("Failed");
             responseListModel.setErrorMsg(ex.getMessage());
