+++ src/main/java/com/example/skooldio/controller/ProductController.java
@@ -1,5 +1,6 @@
 package com.example.skooldio.controller;
 
+import com.example.skooldio.config.security.AuthUtil;
 import com.example.skooldio.constant.UpdateQuantityMode;
 import com.example.skooldio.entity.Product;
 import com.example.skooldio.model.response.ProductModel;
@@ -32,13 +33,18 @@ public class ProductController {
 
     @PostMapping
     @ApiOperation(value = "create product", response = Product.class)
-    public ResponseModel<Product> create(@RequestBody Product entity) {
+    public ResponseModel<Product> create(@RequestHeader String jwtToken, @RequestBody Product entity) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            Product product = service.create(entity);
-            responseModel.setData(product);
-            responseModel.setMsg("Success");
-            responseModel.setErrorMsg(null);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Product product = service.create(entity);
+                responseModel.setData(product);
+                responseModel.setMsg("Success");
+                responseModel.setErrorMsg(null);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -49,15 +55,20 @@ public class ProductController {
 
     @PutMapping("/{id}")
     @ApiOperation(value = "update product by id", response = Product.class)
-    public ResponseModel<Product> update(@PathVariable Long id, @RequestBody Product entity) {
+    public ResponseModel<Product> update(@RequestHeader String jwtToken, @PathVariable Long id, @RequestBody Product entity) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            Product product = service.update(id, entity);
-            responseModel.setMsg("Success");
-            responseModel.setData(product);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Product product = service.update(id, entity);
+                responseModel.setMsg("Success");
+                responseModel.setData(product);
 
-            if (null == product) {
-                responseModel.setErrorMsg(String.format("product id %d not found", id));
+                if (null == product) {
+                    responseModel.setErrorMsg(String.format("product id %d not found", id));
+                }
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
             }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
@@ -68,12 +79,17 @@ public class ProductController {
 
     @PatchMapping(DEDUCT_PRODUCT_ENDPOINT + "/{id}")
     @ApiOperation(value = "deduct product by id", response = Product.class)
-    public ResponseModel<Product> deduct(@PathVariable Long id, @RequestBody int amount) {
+    public ResponseModel<Product> deduct(@RequestHeader String jwtToken, @PathVariable Long id, @RequestBody int amount) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            Product product = service.updateQuantity(id, amount, UpdateQuantityMode.DEDUCT.name());
-            responseModel.setMsg("Success");
-            responseModel.setData(product);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Product product = service.updateQuantity(id, amount, UpdateQuantityMode.DEDUCT.name());
+                responseModel.setMsg("Success");
+                responseModel.setData(product);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -83,12 +99,17 @@ public class ProductController {
 
     @PatchMapping(ADD_PRODUCT_ENDPOINT + "/{id}")
     @ApiOperation(value = "add product by id", response = Product.class)
-    public ResponseModel<Product> add(@PathVariable Long id, @RequestBody int amount) {
+    public ResponseModel<Product> add(@RequestHeader String jwtToken, @PathVariable Long id, @RequestBody int amount) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            Product product = service.updateQuantity(id, amount, UpdateQuantityMode.ADD.name());
-            responseModel.setMsg("Success");
-            responseModel.setData(product);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Product product = service.updateQuantity(id, amount, UpdateQuantityMode.ADD.name());
+                responseModel.setMsg("Success");
+                responseModel.setData(product);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -98,12 +119,17 @@ public class ProductController {
 
     @PatchMapping(DEDUCT_PRODUCT_ENDPOINT)
     @ApiOperation(value = "deduct product list", response = Product.class)
-    public ResponseListModel<ProductModel> deductList(@RequestBody UpdateProductQuantityListModel model) {
+    public ResponseListModel<ProductModel> deductList(@RequestHeader String jwtToken, @RequestBody UpdateProductQuantityListModel model) {
         ResponseListModel<ProductModel> responseListModel = new ResponseListModel<>();
         try {
-            List<ProductModel> products = service.updateQuantityList(model, UpdateQuantityMode.DEDUCT.name());
-            responseListModel.setMsg("Success");
-            responseListModel.setDatas(products);
+            if (AuthUtil.validateToken(jwtToken)) {
+                List<ProductModel> products = service.updateQuantityList(model, UpdateQuantityMode.DEDUCT.name());
+                responseListModel.setMsg("Success");
+                responseListModel.setDatas(products);
+            } else {
+                responseListModel.setMsg("Failed");
+                responseListModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseListModel.setMsg("Failed");
             responseListModel.setErrorMsg(ex.getMessage());
@@ -113,12 +139,17 @@ public class ProductController {
 
     @PatchMapping(ADD_PRODUCT_ENDPOINT)
     @ApiOperation(value = "add product list", response = Product.class)
-    public ResponseListModel<ProductModel> addList(@RequestBody UpdateProductQuantityListModel model) {
+    public ResponseListModel<ProductModel> addList(@RequestHeader String jwtToken, @RequestBody UpdateProductQuantityListModel model) {
         ResponseListModel<ProductModel> responseListModel = new ResponseListModel<>();
         try {
-            List<ProductModel> products = service.updateQuantityList(model, UpdateQuantityMode.ADD.name());
-            responseListModel.setMsg("Success");
-            responseListModel.setDatas(products);
+            if (AuthUtil.validateToken(jwtToken)) {
+                List<ProductModel> products = service.updateQuantityList(model, UpdateQuantityMode.ADD.name());
+                responseListModel.setMsg("Success");
+                responseListModel.setDatas(products);
+            } else {
+                responseListModel.setMsg("Failed");
+                responseListModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseListModel.setMsg("Failed");
             responseListModel.setErrorMsg(ex.getMessage());
@@ -128,18 +159,22 @@ public class ProductController {
 
     @GetMapping("/{id}")
     @ApiOperation(value = "get product by id", response = Product.class)
-    public ResponseModel<Product> getById(@PathVariable Long id) {
+    public ResponseModel<Product> getById(@RequestHeader String jwtToken, @PathVariable Long id) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            Product product = service.getById(id);
-            responseModel.setData(product);
-            responseModel.setMsg("Success");
-            responseModel.setErrorMsg(null);
+            if (AuthUtil.validateToken(jwtToken)) {
+                Product product = service.getById(id);
+                responseModel.setData(product);
+                responseModel.setMsg("Success");
+                responseModel.setErrorMsg(null);
 
-            if (null == product) {
-                responseModel.setErrorMsg(String.format("product id %d not found", id));
+                if (null == product) {
+                    responseModel.setErrorMsg(String.format("product id %d not found", id));
+                }
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
             }
-
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -150,6 +185,7 @@ public class ProductController {
     @GetMapping(LIST_BY_NAME_ENDPOINT)
     @ApiOperation(value = "list product by name", response = Product.class)
     public ResponseListModel<Product> listByName(
+            @RequestHeader String jwtToken,
             @RequestParam String name,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size,
@@ -158,24 +194,28 @@ public class ProductController {
     ) {
         ResponseListModel<Product> responseListModel = new ResponseListModel<>();
         try {
-            List<Product> products = service.listByName(name, page, size, sort, dir);
-            int countAll = service.countByName(name);
-            int count = products.size();
-            int next = 0;
-            if (count >= size) {
-                next = page + size;
-            }
-            if (next >= countAll) {
-                next = 0;
-            }
+            if (AuthUtil.validateToken(jwtToken)) {
+                List<Product> products = service.listByName(name, page, size, sort, dir);
+                int countAll = service.countByName(name);
+                int count = products.size();
+                int next = 0;
+                if (count >= size) {
+                    next = page + size;
+                }
+                if (next >= countAll) {
+                    next = 0;
+                }
 
-            responseListModel.setDatas(products);
-            responseListModel.setAll(countAll);
-            responseListModel.setCount(count);
-            responseListModel.setNext(next);
-            responseListModel.setMsg("Success");
-            responseListModel.setErrorMsg(null);
-
+                responseListModel.setDatas(products);
+                responseListModel.setAll(countAll);
+                responseListModel.setCount(count);
+                responseListModel.setNext(next);
+                responseListModel.setMsg("Success");
+                responseListModel.setErrorMsg(null);
+            } else {
+                responseListModel.setMsg("Failed");
+                responseListModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseListModel.setMsg("Failed");
             responseListModel.setErrorMsg(ex.getMessage());
@@ -185,14 +225,18 @@ public class ProductController {
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "delete product by id", response = Product.class)
-    public ResponseModel<Product> delete(@PathVariable Long id) {
+    public ResponseModel<Product> delete(@RequestHeader String jwtToken, @PathVariable Long id) {
         ResponseModel<Product> responseModel = new ResponseModel<>();
         try {
-            service.deleteById(id);
-            responseModel.setData(null);
-            responseModel.setMsg(String.format("product id %d deleted.", id));
-            responseModel.setErrorMsg(null);
-
+            if (AuthUtil.validateToken(jwtToken)) {
+                service.deleteById(id);
+                responseModel.setData(null);
+                responseModel.setMsg(String.format("product id %d deleted.", id));
+                responseModel.setErrorMsg(null);
+            } else {
+                responseModel.setMsg("Failed");
+                responseModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
+            }
         } catch (Exception ex) {
             responseModel.setMsg("Failed");
             responseModel.setErrorMsg(ex.getMessage());
@@ -203,6 +247,7 @@ public class ProductController {
     @GetMapping
     @ApiOperation(value = "get product as list", response = Product.class)
     public ResponseListModel<Product> listPaging(
+            @RequestHeader String jwtToken,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size,
             @RequestParam(defaultValue = "id") String sort,
@@ -210,24 +255,30 @@ public class ProductController {
     ) {
         ResponseListModel<Product> responseListModel = new ResponseListModel<>();
         try {
-            List<Product> products = service.listPaging(page, size, sort, dir);
+            if (AuthUtil.validateToken(jwtToken)) {
+                List<Product> products = service.listPaging(page, size, sort, dir);
 
-            int countAll = service.countAll();
-            int count = products.size();
-            int next = 0;
-            if (count >= size) {
-                next = page + size;
-            }
-            if (next >= countAll) {
-                next = 0;
+                int countAll = service.countAll();
+                int count = products.size();
+                int next = 0;
+                if (count >= size) {
+                    next = page + size;
+                }
+                if (next >= countAll) {
+                    next = 0;
+                }
+
+                responseListModel.setDatas(products);
+                responseListModel.setAll(countAll);
+                responseListModel.setCount(count);
+                responseListModel.setNext(next);
+                responseListModel.setMsg("Success");
+                responseListModel.setErrorMsg(null);
+            } else {
+                responseListModel.setMsg("Failed");
+                responseListModel.setErrorMsg(AuthUtil.JWT_VERIFY_FAILED);
             }
 
-            responseListModel.setDatas(products);
-            responseListModel.setAll(countAll);
-            responseListModel.setCount(count);
-            responseListModel.setNext(next);
-            responseListModel.setMsg("Success");
-            responseListModel.setErrorMsg(null);
         } catch (Exception ex) {
             responseListModel.setMsg("Failed");
             responseListModel.setErrorMsg(ex.getMessage());
